requirements.txt
request,
pandas,
streamlit,
plotly
